Added configuration for prose
As mentioned in #14, added configuration to prose to allow for an easier
interface when writing blog posts. Excluded unneeded files and folders, setup
a media folder, added the hidden blog field, and create a tags field where a
user could enter in multiple tags.

Could use more specs on what other UI this would
need to be successful.