Created blog section
This includes:
- Blog page with blog excerpts and links with all blog titles on right
- Specific blog page
- All configuration

Styled single blog page

This is a full width page with large text for the blog. Could also make it
have a sidebar.

Had to create a new header to deal with the different header format for the
blog post. Have to re-look at this.

TODO:
- Removed duplicated in header template.
- Possibly add an aside for the sidebar with other blog posts

Got the basic layout going

Before I continue, it would be useful to know what types of other
content to put in these posts. Would we have things like:
- about the author
- author social media links
- social media links? if so, to what